OCM 3.9d
Added improvements to the SDRAM controller using banks by Victor Trucco. Those improvements were done by KdL, that has extrapolated what Victor done for OCM 3.7 releases to the requirements of OCM 3.9.

Removed the hability to turn off debounce on SMX-HB for now. It was causing debounce to be always off. Since that hability has not proven to fix support for devices on joystick port not working, I've just disabled that feature for now, will think whether it will be fully removed or restored / re-tested at a later date.